+++ apps/prairielearn/src/pages/instructorInstanceAdminPublishing/instructorInstanceAdminPublishing.tsx
@@ -8,14 +8,13 @@ import * as error from '@prairielearn/error';
 import { flash } from '@prairielearn/flash';
 import { loadSqlEquiv, queryRows, runInTransactionAsync } from '@prairielearn/postgres';
 import { Hydrate } from '@prairielearn/react/server';
-import { DatetimeLocalStringSchema } from '@prairielearn/zod';
+import { DatetimeLocalStringSchema, parseRequestBody } from '@prairielearn/zod';
 
 import { PageLayout } from '../../components/PageLayout.js';
 import { type AuthzData, assertHasRole } from '../../lib/authz-data-lib.js';
 import { b64EncodeUnicode } from '../../lib/base64-util.js';
 import { extractPageContext } from '../../lib/client/page-context.js';
 import { isRenderableComment } from '../../lib/comments.js';
-import { config } from '../../lib/config.js';
 import { type CourseInstance, CourseInstanceAccessRuleSchema } from '../../lib/db-types.js';
 import { getOriginalHash } from '../../lib/editorUtil.js';
 import { propertyValueWithDefault } from '../../lib/editorUtil.shared.js';
@@ -206,7 +205,6 @@ router.get(
               csrfToken={csrfToken}
               origHash={origHash}
               extensions={publishingExtensions}
-              isDevMode={config.devMode}
             />
           </Hydrate>
         ) : (
@@ -267,23 +265,24 @@ router.post(
         await fs.readFile(infoCourseInstancePath, 'utf8'),
       );
 
-      const parsedBody = z
-        .object({
+      const body = parseRequestBody(
+        req,
+        z.object({
           start_date: z.union([z.literal(''), DatetimeLocalStringSchema]),
           end_date: z.union([z.literal(''), DatetimeLocalStringSchema]),
-        })
-        .parse(req.body);
+        }),
+      );
 
       // Update the publishing settings
       const resolvedPublishing = {
         startDate: propertyValueWithDefault(
           courseInstanceInfo.publishing?.startDate,
-          parsedBody.start_date,
+          body.start_date,
           (v: string) => v === '',
         ),
         endDate: propertyValueWithDefault(
           courseInstanceInfo.publishing?.endDate,
-          parsedBody.end_date,
+          body.end_date,
           (v: string) => v === '',
         ),
       };
@@ -362,11 +361,7 @@ router.post(
           EmailsSchema,
         ),
       });
-      const addExtensionBodyResult = AddExtensionSchema.safeParse(req.body);
-      if (!addExtensionBodyResult.success) {
-        throw new error.HttpStatusError(400, 'Invalid request body');
-      }
-      const body = addExtensionBodyResult.data;
+      const body = parseRequestBody(req, AddExtensionSchema);
 
       const enrollments = (
         await selectUsersAndEnrollmentsByUidsInCourseInstance({
@@ -405,15 +400,12 @@ router.post(
       res.sendStatus(204);
       return;
     } else if (req.body.__action === 'delete_extension') {
-      const deleteExtensionBodyResult = z
-        .object({
+      const body = parseRequestBody(
+        req,
+        z.object({
           extension_id: z.string().trim().min(1),
-        })
-        .safeParse(req.body);
-      if (!deleteExtensionBodyResult.success) {
-        throw new error.HttpStatusError(400, 'Invalid request body');
-      }
-      const body = deleteExtensionBodyResult.data;
+        }),
+      );
 
       const extension = await selectPublishingExtensionById({
         id: body.extension_id,