+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prairielearn/cache
 
+## 2.1.16
+
+### Patch Changes
+
+- Updated dependencies [00945c2]
+- Updated dependencies [9757f58]
+  - @prairielearn/logger@4.0.0
+
+## 2.1.15
+
+### Patch Changes
+
+- df28e68: Bump dependencies
+
 ## 2.1.14
 
 ### Patch Changes