+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/postgres-tools
 
+## 3.0.13
+
+### Patch Changes
+
+- 057d37b: Update dependency on typescript-cp
+
+## 3.0.12
+
+### Patch Changes
+
+- df28e68: Bump dependencies
+
 ## 3.0.11
 
 ### Patch Changes